+++ resources/views/clientes/create.blade.php
@@ -20,10 +20,5 @@
             </h2>
         </div>
     </x-slot>
-    <x-form-clients action="{{ route('clientes.store') }}" method="POST" :company="$company" />
-
-    @push('scripts')
-        @vite(['resources/js/validations-client.js'])
-        @vite(['resources/js/search-company.js'])
-    @endpush
+    <x-form-clients action="{{ route('clientes.store') }}" method="POST" :company="$company" :client="$client" />
 </x-app-layout>